app_python/requirements.txt
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
sqlalchemy==2.0.30
requests==2.32.3
schedule==1.2.2
python-dotenv==1.0.1
pytest==8.2.2
